requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
markdown==3.5.1
Pygments==2.17.2
gunicorn==21.2.0
python-docx==1.1.0
zhipuai>=2.1.5
openai
sniffio
httpx
pydantic
requests
openpyxl>=3.1.0
flask-login>=0.6.3
